pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Guilherme Halter", email = "guisam.halter@gmail.com"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=67.6.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
